+++ sshmitm/forwarders/ssh.py
@@ -39,6 +39,9 @@ class SSHForwarder(SSHBaseForwarder):
     def forward(self) -> None:
         time.sleep(0.1)
 
+        # store the remote channel to the ssh server in the current session
+        self.session.ssh_remote_channel = self.server_channel
+
         if self.session.ssh_pty_kwargs:
             self.server_channel.get_pty(**self.session.ssh_pty_kwargs)
         self.server_channel.invoke_shell()